backend/mosupisi-llm-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
llama-cpp-python>=0.2.0
pydantic>=2.0.0